--- src/App.css
@@ -1,6 +1,0 @@
-* {
-    padding: 0px;
-    margin: 0px;
-    box-sizing: border-box;
-}
-
