+++ src/App.module.css
@@ -1,0 +1,20 @@
+* {
+    padding: 0px;
+    margin: 0px;
+    box-sizing: border-box;
+}
+
+.footer {
+    background-color: #000;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    color: white;
+    height: 60px;
+}
+.footerText{
+     font-size: 18px; 
+    font-weight: 500;
+    font-family: sans-serif;
+}
