+++ src/components/Cart.module.css
@@ -1,6 +1,13 @@
+* {
+  padding: 0px;
+  margin: 0px;
+  box-sizing: border-box;
+}
+
 ul {
   list-style-type: none;
 }
+
 li {
   padding: 10px;
 }
@@ -11,31 +18,82 @@ li {
   padding: 10px;
   cursor: pointer;
 }
+
 .cartContainer {
-  border: 1px solid black;
-  border-radius: 10px;
-  margin: 20px;
-  height: 400px;
-  padding: 10px;
-  overflow: scroll;
-  overflow-x: hidden;
+  background-color: grey;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  /* width: 100vw; */
 }
 
 .cartItem {
-  display: flex;
-  justify-content: space-between;
-  border: 1px solid black;
-  border-radius: 10px;
-  padding: 20px;
+  width: 60vw;
+  height: 40vh;
   font-size: 20px;
   font-weight: bold;
-  background-color: #2f3c7e;
+  background-color: pink;
   color: #fbeaeb;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  /* align-items: center;  */
+  border-radius: 5px;
+  cursor: pointer;
 }
+
+.text {
+  font-size: .875rem;
+  line-height: 1.25rem;
+
+}
+
 .cartActions {
   color: white;
   font-size: 20px;
   width: 50px;
   border: none;
   background: rgb(10, 7, 24);
+}
+
+/* .totalPrice {
+  text-align: right;
+  margin-right: 50px;
+  margin-bottom: 20px;
+  margin-top: 1rem;
+  font-size: 20px;
+} */
+
+.orderBtn {
+  width: 200px;
+  height: 50px;
+  background: orangered;
+  border: none;
+  cursor: pointer;
+}
+
+.orderBtn:hover {
+  background: rgb(89, 0, 190);
+}
+
+.imageItem {
+  background-color: black;
+  width: 20%;
+  height: auto;
+}
+
+.item {
+  background-color: darkcyan;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  width: 50%;
+  height: 30%;
+}
+
+.itemsList {
+  background-color: darkmagenta;
+
 }
