+++ src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
-import './index.css'
+import styles from'./index.module.css'
 import { Provider } from 'react-redux'
 import store from './store/index.jsx'
 import { BrowserRouter } from "react-router-dom"
